src/test/javascript/spec/app/entities/caisse-maladie/caisse-maladie.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { CaisseMaladieService } from 'app/entities/caisse-maladie/caisse-maladie.service';
import { ICaisseMaladie, CaisseMaladie } from 'app/shared/model/caisse-maladie.model';

describe('Service Tests', () => {
  describe('CaisseMaladie Service', () => {
    let injector: TestBed;
    let service: CaisseMaladieService;
    let httpMock: HttpTestingController;
    let elemDefault: ICaisseMaladie;
    let expectedResult: ICaisseMaladie | ICaisseMaladie[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(CaisseMaladieService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new CaisseMaladie(0, 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a CaisseMaladie', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new CaisseMaladie()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a CaisseMaladie', () => {
        const returnedFromService = Object.assign(
          {
            numeroCaisseMaladie: 'BBBBBB',
            nom: 'BBBBBB',
            adresse: 'BBBBBB',
            npaLocalite: 'BBBBBB',
            telephone: 'BBBBBB',
            fax: 'BBBBBB',
            email: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of CaisseMaladie', () => {
        const returnedFromService = Object.assign(
          {
            numeroCaisseMaladie: 'BBBBBB',
            nom: 'BBBBBB',
            adresse: 'BBBBBB',
            npaLocalite: 'BBBBBB',
            telephone: 'BBBBBB',
            fax: 'BBBBBB',
            email: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a CaisseMaladie', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
